# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for application",
        "title": "Test task Swartway",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/employee/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Создание работника",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createEmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/employee/delete/{employeeId}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Удаление работника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор работника",
                        "name": "employeeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseOk"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/employee/list/department/{depName}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employees"
                ],
                "summary": "Получение работников по отделу",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название отдела",
                        "name": "depName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseWithDetails-handler_listEmployeesByDepartmentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/employee/list/{companyId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Получение работников по id компании",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор компании",
                        "name": "companyId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseWithDetails-handler_getListEmployeesByCompanyIdResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/employee/update/{employeeId}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Изменение данных работника",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор пользователя",
                        "name": "employeeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "da",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.updateEmployeeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseWithDetails-handler_updateEmployeeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    },
                    "409": {
                        "description": "Already exists",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/ping": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Служебные"
                ],
                "summary": "Пинг сервиса",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/thttp.ResponseWithDetails-string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.createEmployeeRequest": {
            "type": "object",
            "required": [
                "company_id",
                "department_name",
                "department_phone",
                "id",
                "name",
                "passport_number",
                "passport_type",
                "phone",
                "surname"
            ],
            "properties": {
                "company_id": {
                    "type": "integer",
                    "example": 1
                },
                "department_name": {
                    "type": "string",
                    "example": "First Department"
                },
                "department_phone": {
                    "type": "string",
                    "example": "+7(987)1112233"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Ivan"
                },
                "passport_number": {
                    "type": "string",
                    "example": "1122 112233"
                },
                "passport_type": {
                    "type": "string",
                    "example": "Russian passport"
                },
                "phone": {
                    "type": "string",
                    "example": "+7(987)6667788"
                },
                "surname": {
                    "type": "string",
                    "example": "Ivanov"
                }
            }
        },
        "handler.getListEmployeesByCompanyIdElement": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "department_name": {
                    "type": "string"
                },
                "department_phone": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "passport_type": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "handler.listEmployeesByDepartmentElement": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "department_name": {
                    "type": "string"
                },
                "department_phone": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "passport_type": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "handler.updateEmployeeRequest": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "department_name": {
                    "type": "string"
                },
                "department_phone": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "passport_type": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "handler.updateEmployeeResponse": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "integer"
                },
                "department_name": {
                    "type": "string"
                },
                "department_phone": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "passport_number": {
                    "type": "string"
                },
                "passport_type": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "thttp.ResponseError": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {}
            }
        },
        "thttp.ResponseOk": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {}
            }
        },
        "thttp.ResponseWithDetails-handler_getListEmployeesByCompanyIdResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.getListEmployeesByCompanyIdElement"
                    }
                }
            }
        },
        "thttp.ResponseWithDetails-handler_listEmployeesByDepartmentResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.listEmployeesByDepartmentElement"
                    }
                }
            }
        },
        "thttp.ResponseWithDetails-handler_updateEmployeeResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "$ref": "#/definitions/handler.updateEmployeeResponse"
                }
            }
        },
        "thttp.ResponseWithDetails-string": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "error_code": {
                    "type": "integer"
                },
                "ok": {
                    "type": "boolean"
                },
                "result": {
                    "type": "string"
                }
            }
        }
    }
}